BoundingBox class
