+++ attributes/AttributesForInfo.js
@@ -1,10 +1,9 @@
 var conn = require('../common/conn');
 var Promise = require('promise');
-var extent = require('geojson-extent');
-var wellknown = require('wellknown');
 var logger = require('../common/Logger').applicationWideLogger;
 
 let Attributes = require('./Attributes');
+let BoundingBox = require('../custom_features/BoundingBox');
 
 var MongoAttribute = require('../attributes/MongoAttribute');
 var NumericAttribute = require('../attributes/NumericAttribute');
@@ -73,7 +72,7 @@ class AttributesForInfo extends Attributes {
           dataViews.map(view => {
             if (view.rows.length > 0){
               let info = view.rows[0];
-              let wgsExtent = this.getExtentFromWkt(info.geomwgs);
+              let wgsExtent = new BoundingBox().getExtentFromWkt(info.geomwgs);
                 logger.info(`AttributesForInfo#statisticAttributes extent: ${wgsExtent}`);
               data = {
                 gid: info.gid,
@@ -103,11 +102,6 @@ class AttributesForInfo extends Attributes {
       }, conn.getMongoDb()).read().then(baseLayers => {
           return baseLayers;
       })
-  }
-
-  getExtentFromWkt(geometry){
-      let json = wellknown(geometry);
-      return extent(json);
   }
 }
